# openapi/private.yaml
openapi: 3.0.2
info:
  title: "Bangumi Private API"
  version: ""
  description: "`next.bgm.tv` 域名不支持跨域，"
servers:
  - url: https://next.bgm.tv
    description: Production server
...
paths:
...
  "/p/me":
    get:
      tags:
        - Auth
      summary: 获取当前用户
      operationId: getCurrentUser
      security:
        - CookiesSession: []
      responses:
        "200":
          description: "返回当前用户"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "401":
          description: 用户未登录或者 session 已失效
...
components:
  schemas:
    User:
      $ref: "./components/user.yaml"
...
  securitySchemes:
    CookiesSession:
      name: sessionID
      type: apiKey
      in: cookie